Fixed code that was leaving zombie child processes
This has probably been a problem all along, this daemon was never
waiting for child processes to finish and was never itself exiting
(because.. it's a daemon). Now waiting for these (hopefully very fast)
child processes started with Process.runCommand
In the future, if necessary, can fork another process to wait for them
in the Haskell code.